requirements.txt
chardet
psycopg2-binary
python-dotenv